Address some of the PR comments
