+++ lib/providers/filesystem/index.js
@@ -20,11 +20,11 @@ function FileSystemProvider(options) {
   this.root = options.root;
   var exists = fs.existsSync(this.root);
   if (!exists) {
-    throw new Error('Path does not exist: ' + this.root);
+    throw new Error('FileSystemProvider: Path does not exist: ' + this.root);
   }
   var stat = fs.statSync(this.root);
   if (!stat.isDirectory()) {
-    throw new Error('Invalid directory: ' + this.root);
+    throw new Error('FileSystemProvider: Invalid directory: ' + this.root);
   }
 }
 
@@ -34,7 +34,7 @@ function validateName(name, cb) {
   if (!name) {
     cb && process.nextTick(cb.bind(null, new Error('Invalid name: ' + name)));
     if (!cb) {
-      console.error('Invalid name: ', name);
+      console.error('FileSystemProvider: Invalid name: ', name);
     }
     return false;
   }
@@ -42,9 +42,10 @@ function validateName(name, cb) {
   if (match && match.index === 0 && match[0].length === name.length) {
     return true;
   } else {
-    cb && process.nextTick(cb.bind(null, new Error('Invalid name: ' + name)));
+    cb && process.nextTick(cb.bind(null,
+      new Error('FileSystemProvider: Invalid name: ' + name)));
     if (!cb) {
-      console.error('Invalid name: ', name);
+      console.error('FileSystemProvider: Invalid name: ', name);
     }
     return false;
   }
@@ -57,7 +58,7 @@ FileSystemProvider.prototype.getContainers = function (cb) {
     var containers = [];
     var tasks = [];
     files.forEach(function (f) {
-      tasks.push(fs.stat.bind(null, path.join(self.root, f)));
+      tasks.push(fs.stat.bind(fs, path.join(self.root, f)));
     });
     async.parallel(tasks, function (err, stats) {
       if (err) {
@@ -95,7 +96,7 @@ FileSystemProvider.prototype.destroyContainer = function (containerName, cb) {
   fs.readdir(dir, function (err, files) {
     var tasks = [];
     files.forEach(function (f) {
-      tasks.push(fs.unlink.bind(null, path.join(dir, f)));
+      tasks.push(fs.unlink.bind(fs, path.join(dir, f)));
     });
     async.parallel(tasks, function (err) {
       if (err) {
@@ -132,9 +133,10 @@ FileSystemProvider.prototype.upload = function (options, cb) {
   if (!validateName(file, cb)) return;
   var filePath = path.join(this.root, container, file);
 
-  var fileOpts = {flags: 'w+',
-    encoding: null,
-    mode: 0666 };
+  var fileOpts = {flags: options.flags || 'w+',
+    encoding: options.encoding || null,
+    mode: options.mode || 0666
+  };
 
   try {
     return fs.createWriteStream(filePath, fileOpts);
@@ -173,7 +175,7 @@ FileSystemProvider.prototype.getFiles = function (container, download, cb) {
     var files = [];
     var tasks = [];
     entries.forEach(function (f) {
-      tasks.push(fs.stat.bind(null, path.join(dir, f)));
+      tasks.push(fs.stat.bind(fs, path.join(dir, f)));
     });
     async.parallel(tasks, function (err, stats) {
       if (err) {
